+++ corehq/couchapps/tests/test_dbaccessors.py
@@ -2,8 +2,7 @@ from django.core.files.uploadedfile import UploadedFile
 
 from corehq.apps.receiverwrapper import submit_form_locally
 from corehq.couchapps.dbaccessors import get_attachment_size_by_domain, get_attachment_size_by_domain_app_id_xmlns
-from corehq.form_processor.interfaces.processor import FormProcessorInterface
-from corehq.form_processor.test_utils import FormProcessorTestUtils
+from corehq.form_processor.tests.utils import FormProcessorTestUtils
 from django.test import TestCase
 from dimagi.utils.make_uuid import random_hex
