+++ corehq/ex-submodules/casexml/apps/case/tests/test_strict_datetimes.py
@@ -6,7 +6,7 @@ from casexml.apps.case.util import make_form_from_case_blocks
 from casexml.apps.case.xform import process_cases_with_casedb
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from corehq.form_processor.parsers.form import process_xform
-from corehq.form_processor.test_utils import run_with_all_backends
+from corehq.form_processor.tests.utils import run_with_all_backends
 
 
 CASE_ID = 'a0cd5e6c5fb84695a4f729d3b1996a93'
