+++ corehq/ex-submodules/casexml/apps/case/tests/util.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from xml.etree import ElementTree
 from corehq.apps.domain.models import Domain
 from corehq.apps.receiverwrapper import submit_form_locally
-from corehq.form_processor.test_utils import FormProcessorTestUtils
+from corehq.form_processor.tests.utils import FormProcessorTestUtils
 from corehq.util.test_utils import unit_testing_only
 
 from dimagi.utils.dates import utcnow_sans_milliseconds
