+++ corehq/ex-submodules/couchforms/tests/test_devicelogs.py
@@ -4,7 +4,7 @@ from django.test import TestCase
 from corehq.apps.receiverwrapper import submit_form_locally
 from corehq.util.test_utils import TestFileMixin
 from corehq.form_processor.interfaces.processor import FormProcessorInterface
-from corehq.form_processor.test_utils import run_with_all_backends
+from corehq.form_processor.tests.utils import run_with_all_backends
 from phonelog.models import UserEntry, DeviceReportEntry
